refactor(server): extract CLI/startup into cli.py (#183 part A)
server.py mixed tool/resource/prompt registration with CLI parsing and
startup wiring. Move `_cmd_install_skill` and `main()` verbatim into a new
`cli.py`; server.py keeps the `_session`/`_has_library` globals (read by every
tool closure) and gains a small `configure(session, has_library)` setter that
cli.main() calls. Entry point flips `server:main` -> `cli:main`.

No behavior change. server.py drops 891 -> 741 lines and is now registration +
the session-wiring setter. The live MCP tests launch the `build123d-mcp`
console script, so the entry-point change is exercised end-to-end; the
install-skill subcommand path is covered by test_install_skill.py.

Part A of two for #183 (Part B extracts render marshalling). Does not close
the issue on its own.
